+++ lib/cocoapods/validator.rb
@@ -197,6 +197,7 @@ module Pod
     #
     def perform_extensive_analysis(spec)
       validate_homepage(spec)
+      validate_screenshots(spec)
 
       spec.available_platforms.each do |platform|
         UI.message "\n\n#{spec} - Analyzing on #{platform} platform.".green.reversed
@@ -222,22 +223,60 @@ module Pod
     attr_accessor :consumer
     attr_accessor :subspec_name
 
-    # Performs validations related to the `homepage` attribute.
+    MAX_HTTP_REDIRECTS = 3
+
+    # Performs validation of a URL
     #
-    def validate_homepage(spec)
+    def validate_url(url)
       require 'rest'
-      homepage = spec.homepage
-      return unless homepage
 
       begin
-        resp = ::REST.head(homepage)
+        redirects = 0
+        resp = nil
+        loop do
+          resp = ::REST.head(url)
+
+          if resp.status_code >= 400
+            resp = ::REST.get(url)
+          end
+
+          if [301, 302, 303, 307, 308].include? resp.status_code
+            url = resp.headers['location'].first
+            redirects += 1
+          else
+            break
+          end
+
+          break unless redirects < MAX_HTTP_REDIRECTS
+        end
       rescue
-        warning "There was a problem validating the homepage."
+        warning "There was a problem validating the URL #{url}."
         resp = nil
       end
 
       if resp && !resp.success?
-        warning "The homepage is not reachable."
+        warning "The URL (#{url}) is not reachable."
+      end
+
+      resp
+    end
+
+    # Performs validations related to the `homepage` attribute.
+    #
+    def validate_homepage(spec)
+      if spec.homepage
+        validate_url(spec.homepage)
+      end
+    end
+
+    # Performs validation related to the `screenshots` attribute.
+    #
+    def validate_screenshots(spec)
+      spec.screenshots.compact.each do |screenshot|
+        request = validate_url(screenshot)
+        if request && !(request.headers['content-type'] && request.headers['content-type'].first =~ /image\/.*/i)
+          warning "The screenshot #{screenshot} is not a valid image."
+        end
       end
     end