+++ spec/unit/validator_spec.rb
@@ -75,6 +75,7 @@ module Pod
         validator = Validator.new(file)
         validator.quick = true
         validator.stubs(:validate_homepage)
+        validator.stubs(:validate_screenshots)
         validator.validate
         validator.validation_dir.should.be == Pathname.new("/private/tmp/CocoaPods/Lint")
       end
@@ -97,16 +98,79 @@ module Pod
 
         it "checks if the homepage is valid" do
           WebMock::API.stub_request(:head, /not-found/).to_return(:status => 404)
+          WebMock::API.stub_request(:get, /not-found/).to_return(:status => 404)
           Specification.any_instance.stubs(:homepage).returns('http://banana-corp.local/not-found/')
           @sut.validate
-          @sut.results.map(&:to_s).first.should.match /The homepage is not reachable/
+          @sut.results.map(&:to_s).first.should.match /The URL (.*) is not reachable/
         end
 
         it "indicates if it was not able to validate the homepage" do
           WebMock::API.stub_request(:head, 'banana-corp.local').to_raise(SocketError)
           Specification.any_instance.stubs(:homepage).returns('http://banana-corp.local/')
           @sut.validate
-          @sut.results.map(&:to_s).first.should.match /There was a problem validating the homepage/
+          @sut.results.map(&:to_s).first.should.match /There was a problem validating the URL/
+        end
+
+        it "does not fail if the homepage redirects" do
+          WebMock::API.stub_request(:head, /redirect/).to_return(
+            :status => 301, :headers => { 'Location' => 'http://banana-corp.local/found/' } )
+          WebMock::API.stub_request(:head, /found/).to_return( :status => 200 )
+          Specification.any_instance.stubs(:homepage).returns('http://banana-corp.local/redirect/')
+          @sut.validate
+          @sut.results.length.should.equal 0
+        end
+
+        it "does not fail if the homepage does not support HEAD" do
+          WebMock::API.stub_request(:head, /page/).to_return( :status => 405 )
+          WebMock::API.stub_request(:get, /page/).to_return( :status => 200 )
+          Specification.any_instance.stubs(:homepage).returns('http://banana-corp.local/page/')
+          @sut.validate
+          @sut.results.length.should.equal 0
+        end
+
+        it "does not fail if the homepage errors on HEAD" do
+          WebMock::API.stub_request(:head, /page/).to_return( :status => 500 )
+          WebMock::API.stub_request(:get, /page/).to_return( :status => 200 )
+          Specification.any_instance.stubs(:homepage).returns('http://banana-corp.local/page/')
+          @sut.validate
+          @sut.results.length.should.equal 0
+        end
+
+        it "does not follow redirects infinitely" do
+          WebMock::API.stub_request(:head, /redirect/).to_return(
+            :status => 301,
+            :headers => { 'Location' => 'http://banana-corp.local/redirect/' } )
+          Specification.any_instance.stubs(:homepage).returns(
+            'http://banana-corp.local/redirect/')
+          @sut.validate
+          @sut.results.map(&:to_s).first.should.match /The URL \(.*\) is not reachable/
+        end
+      end
+
+      describe "Screenshot validation" do
+        require 'webmock'
+
+        before do
+          @sut = Validator.new(podspec_path)
+          @sut.stubs(:install_pod)
+          @sut.stubs(:build_pod)
+          @sut.stubs(:check_file_patterns)
+          @sut.stubs(:tear_down_validation_environment)
+          @sut.stubs(:validate_homepage)
+          WebMock::API.stub_request(:head, 'banana-corp.local/valid-image.png').to_return(:status => 200, :headers => { 'Content-Type' => 'image/png' })
+        end
+
+        it "checks if the screenshots are valid" do
+          Specification.any_instance.stubs(:screenshots).returns(['http://banana-corp.local/valid-image.png'])
+          @sut.validate
+          @sut.results.should.be.empty?
+        end
+
+        it "should fail if any of the screenshots URLS do not return an image" do
+          WebMock::API.stub_request(:head, 'banana-corp.local/').to_return(:status => 200)
+          Specification.any_instance.stubs(:screenshots).returns(['http://banana-corp.local/valid-image.png', 'http://banana-corp.local/'])
+          @sut.validate
+          @sut.results.map(&:to_s).first.should.match /The screenshot .* is not a valid image/
         end
       end
 
@@ -114,6 +178,7 @@ module Pod
         file = write_podspec(stub_podspec)
         sut = Validator.new(file)
         sut.stubs(:validate_homepage)
+        sut.stubs(:validate_screenshots)
         sut.no_clean = true
         sut.validate
         sut.validation_dir.should.exist
@@ -123,6 +188,7 @@ module Pod
         file = write_podspec(stub_podspec)
         sut = Validator.new(file)
         sut.stubs(:validate_homepage)
+        sut.stubs(:validate_screenshots)
         sut.expects(:install_pod).twice
         sut.expects(:build_pod).twice
         sut.expects(:check_file_patterns).twice
@@ -132,6 +198,7 @@ module Pod
       it "uses the deployment target of the specification" do
         sut = Validator.new(podspec_path)
         sut.stubs(:validate_homepage)
+        sut.stubs(:validate_screenshots)
         podfile = sut.send(:podfile_from_spec, :ios, '5.0')
         dependency = podfile.target_definitions['Pods'].dependencies.first
         dependency.external_source.has_key?(:podspec).should.be.true
@@ -140,6 +207,7 @@ module Pod
       it "respects the local option" do
         sut = Validator.new(podspec_path)
         sut.stubs(:validate_homepage)
+        sut.stubs(:validate_screenshots)
         podfile = sut.send(:podfile_from_spec, :ios, '5.0')
         deployment_target = podfile.target_definitions['Pods'].platform.deployment_target
         deployment_target.to_s.should == "5.0"
@@ -150,6 +218,7 @@ module Pod
         sut.stubs(:check_file_patterns)
         sut.stubs(:xcodebuild).returns("file.m:1:1: warning: direct access to objective-c's isa is deprecated")
         sut.stubs(:validate_homepage)
+        sut.stubs(:validate_screenshots)
         sut.validate
         first = sut.results.map(&:to_s).first
         first.should.include "[xcodebuild]"
@@ -161,6 +230,7 @@ module Pod
         sut = Validator.new(file)
         sut.stubs(:build_pod)
         sut.stubs(:validate_homepage)
+        sut.stubs(:validate_screenshots)
         sut.validate
         sut.results.map(&:to_s).first.should.match /source_files.*did not match/
         sut.result_type.should == :error
@@ -175,6 +245,7 @@ module Pod
         spec = Specification.from_file(file)
         sut = Validator.new(spec)
         sut.stubs(:validate_homepage)
+        sut.stubs(:validate_screenshots)
         sut.stubs(:build_pod)
         sut.validate
         sut.validated?.should.be.true